Rename destruct module to destructuring
